Integer Order Functions

Mathieu.ce — Function

ce(m,q,z)

Compute the cosine-elliptic Mathieu function $\text{ce}_m(q,z)$. m is the order of the solution and must be an integer $m=0,1,2,3…$.

Pass a vector, e.g. m=0:4 and/or z=LinRange(0,π,101), to compute the function for multiple orders and variables efficiently.

Mathieu.se — Function

se(m,q,z)

Compute the sine-elliptic Mathieu function $\text{se}_m(q,z)$. m is the order of the solution and must be an integer, excluding 0, $m=1,2,3,4…$.

Pass a vector, e.g. m=1:5 and/or z=LinRange(0,π,101), to compute the function for multiple orders and variables efficiently.