Integer Order Characteristic Values

Mathieu.a — Function

a(m,q)

Compute the characteristic value $a_m(q)$ corresponding to an even (cosine-elliptic) solution to the angular Mathieu equation. m is the order of the solution and must be an integer $m=0,1,2,3…$.

Pass a vector, e.g. m=0:4, to compute multiple characteristic values efficiently.

Mathieu.aa — Function

aa(n,q)

Compute the characteristic value $a_{2n+1}(q)$ corresponding to an even (cosine-elliptic), π-antiperiodic (2π-periodic) solution to the angular Mathieu equation, for $n=0,1,2,3,…$.

Pass a vector, e.g. n=0:4, to compute multiple characteristic values efficiently.

Mathieu.ap — Function

ap(n,q)

Compute the characteristic value $a_{2n}(q)$ corresponding to an even (cosine-elliptic), π-periodic solution to the angular Mathieu equation, for $n=0,1,2,3,…$.

Pass a vector, e.g. n=0:4, to compute multiple characteristic values efficiently.

Mathieu.b — Function

b(m,q)

Compute the characteristic value $b_m(q)$ corresponding to an odd (sine-elliptic) solution to the angular Mathieu equation. m is the order of the solution and must be an integer, excluding 0, $m=1,2,3,4…$.

Pass a vector, e.g. m=1:5, to compute multiple characteristic values efficiently.

Mathieu.ba — Function

ba(n,q)

Compute the characteristic value $b_{2n+1}(q)$ corresponding to an odd (sine-elliptic), π-antiperiodic (2π-periodic) solution to the angular Mathieu equation, for $n=0,1,2,3,…$.

Pass a vector, e.g. n=0:4, to compute multiple characteristic values efficiently.

Mathieu.bp — Function

bp(n,q)

Compute the characteristic value $b_{2n+2}(q)$ corresponding to an odd (sine-elliptic), π-periodic solution to the angular Mathieu equation, for $n=0,1,2,3,…$.

Pass a vector, e.g. n=0:4, to compute multiple characteristic values efficiently.

Mathieu.char_aper — Function

char_aper(m,q)

Compute the characteristic value $a_{m+1}(q)$ for even m, or $b_{m}(q)$ for odd m. These characteristic values correspond respectively to even and odd π-antiperiodic (2π-periodic) solutions to the angular Mathieu equation, for $n=0,1,2,3,…$.

m	value

0	a_1

1	b_1

2	a_3

3	b_3

4	a_5

⋮	⋮

Pass a vector, e.g. m=0:4, to compute multiple characteristic values efficiently.

Mathieu.char_per — Function

char_per(m,q)

Compute the characteristic value $a_{m}(q)$ for even m, or $b_{m+1}(q)$ for odd m. These characteristic values correspond respectively to even and odd π-periodic solutions to the angular Mathieu equation, for $n=0,1,2,3,…$.

m	value

0	a_0

1	b_2

2	a_2

3	b_4

4	a_4

⋮	⋮

Pass a vector, e.g. m=0:4, to compute multiple characteristic values efficiently.